ClientOrders Tests & Unternehmen Mapping
Client Tests werden abgeschloßen
Outstanding wird gut funktioniert
Unternehmen Mapping wird mit keinem Test abgeschloßen
Attribute mit "Nullable" wird korrigiert
Die Summe von den Bestellungen wird in dem Entität berechnet

+++ haegerp/src/test/java/org/haegerp/entity/Article_ArticleCategoryTest.java
@@ -301,7 +301,7 @@ public class Article_ArticleCategoryTest extends TestCase {
 	        article.setSizeH(Float.parseFloat(Config.getProperty("INSERT_A_SIZEH_F")));
 	        article.setSizeL(Float.parseFloat(Config.getProperty("INSERT_A_SIZEL_F")));
 	        article.setSizeW(Float.parseFloat(Config.getProperty("INSERT_A_SIZEW_F")));
-	        article.setStock(Integer.parseInt(Config.getProperty("INSERT_A_STOCK_F")));
+	        article.setStock(Long.parseLong(Config.getProperty("INSERT_A_STOCK_F")));
 	        
 	        article = articleRepo.performNew(article);
